# api/openapi.yaml
openapi: 3.1.1
info:
  title: Sudoku HTTP API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  summary: Authoritative Sudoku game sessions over HTTP
  description: |
    The Sudoku HTTP API exposes the existing game engine to local tools and
    separately deployed clients. The server is authoritative for state,
    revisions, validation, and recovery.

    The server listens on loopback by default. Operators that configure a
    non-loopback listener must also configure bearer authentication. Clients
    should send a bearer credential whenever their deployment requires one.
servers:
  - url: http://127.0.0.1:8080
    description: Default loopback deployment
...
paths:
...
  /api/v1/sessions:
    get:
      tags: [Sessions]
      operationId: listSessions
      summary: List active and recovered sessions
      description: Returns bounded summaries and never returns session payloads.
      responses:
        "200":
          description: Session summaries ordered by most recently updated first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Sessions]
      operationId: createSession
      summary: Create a session from a difficulty or puzzle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
            examples:
              byDifficulty:
                summary: Generate or load an easy puzzle
                value:
                  source:
                    kind: difficulty
                    difficulty: easy
              byPuzzle:
                summary: Start from a supplied puzzle
                value:
                  source:
                    kind: puzzle
                    puzzle: "..3.2.6..9..3.5..1..18.64....81.29..7.......8..67.82....26.95..8..2.3..9..5.1.3.."
      responses:
        "201":
          description: The session was created at revision zero.
          headers:
            Location:
              description: Relative URL of the new session.
              schema:
                type: string
                pattern: ^/api/v1/sessions/[A-Za-z0-9_-]+$
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/import:
    post:
      tags: [Sessions]
      operationId: importSession
      summary: Import a serialized game session
      description: |
        Restores a bounded versioned session document. The request body is the
        session document itself; the API never accepts a host filesystem path.
      requestBody:
        required: true
        content:
          application/vnd.sudoku.session+json:
            schema:
              type: string
              format: binary
              contentMediaType: application/json
              maxLength: 1048576
      responses:
        "201":
          description: The imported session was validated and made durable.
          headers:
            Location:
              description: Relative URL of the imported session.
              schema:
                type: string
                pattern: ^/api/v1/sessions/[A-Za-z0-9_-]+$
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [Gameplay]
      operationId: getSession
      summary: Read the authoritative session state
      responses:
        "200":
          description: The current session revision and detached snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Sessions]
      operationId: deleteSession
      summary: Discard a session and its recovery record
      responses:
        "204":
          description: The session and only its recovery record were deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/hint:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [Gameplay]
      operationId: previewHint
      summary: Preview a hint without mutating the session
      responses:
        "200":
          description: A hint for the current revision. The revision is unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HintPreview"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/export:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [Sessions]
      operationId: exportSession
      summary: Export the versioned serialized session
      description: Returns the persistence document as bounded bytes, not a host path.
      responses:
        "200":
          description: The versioned session document.
          headers:
            Content-Disposition:
              description: Suggested download filename.
              schema:
                type: string
                example: attachment; filename="sudoku-session.json"
          content:
            application/vnd.sudoku.session+json:
              schema:
                type: string
                format: binary
                contentMediaType: application/json
                maxLength: 1048576
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/sessions/{sessionId}/actions:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [Gameplay]
      operationId: applyAction
      summary: Apply one typed game action
      description: |
        The expected revision provides optimistic concurrency control. Every
        accepted mutation increments the revision exactly once. Rejected and
        stale actions leave state and revision unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionRequest"
            examples:
              setValue:
                value:
                  expected_revision: 0
                  kind: set-value
                  row: 1
                  column: 1
                  value: 4
              undo:
                value:
                  expected_revision: 3
                  kind: undo
      responses:
        "200":
          description: The action was accepted and durably persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The expected revision is stale.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevisionConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: Required whenever the server is configured for authenticated access.
...
  responses:
    BadRequest:
      description: The request is malformed or violates the transport contract.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: invalid-request
              message: source must select exactly one supported kind
              field: source
    Unauthorized:
      description: A required bearer credential is missing or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: unauthorized
              message: bearer authentication is required
...
    PayloadTooLarge:
      description: The request exceeds the configured body limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: payload-too-large
              message: request body exceeds the maximum size
...
    InternalError:
      description: An unexpected or durability failure occurred without leaking internals.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal-error
              message: the request could not be completed
  schemas:
...
    Revision:
      type: integer
      format: int64
      minimum: 0
...
    Session:
      type: object
      additionalProperties: false
      required: [id, revision, snapshot]
      properties:
        id:
          $ref: "#/components/schemas/SessionId"
        revision:
          $ref: "#/components/schemas/Revision"
        snapshot:
          $ref: "#/components/schemas/Snapshot"
...
    SessionList:
      type: object
      additionalProperties: false
      required: [sessions]
      properties:
        sessions:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/SessionSummary"
    CreateSessionRequest:
      type: object
      additionalProperties: false
      required: [source]
      properties:
        source:
          oneOf:
            - $ref: "#/components/schemas/DifficultySource"
            - $ref: "#/components/schemas/PuzzleSource"
          discriminator:
            propertyName: kind
            mapping:
              difficulty: "#/components/schemas/DifficultySource"
              puzzle: "#/components/schemas/PuzzleSource"
...
security:
  - {}
  - bearerAuth: []